# packages/api-client/src/generated/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type UserDto = {
    id: number;
    name: string;
    email: string;
    createdAt: string;
    updatedAt: string;
};

export type UsersControllerFindAllResponse = (Array<UserDto>);

export type UsersControllerFindAllError = unknown;

export type UsersControllerFindOneData = {
    path: {
        id: number;
    };
};

export type UsersControllerFindOneResponse = (UserDto);

export type UsersControllerFindOneError = (unknown);